code review and improve compose file

+++ server/src/main/java/com/eestienergia/fundraiser/domain/Basket.java
@@ -9,14 +9,14 @@ import java.util.Map;
 @Getter
 public class Basket {
 
-    private final Map<Product, Long> items = new HashMap<>();
+    private final Map<Product, Integer> items = new HashMap<>();
     private BigDecimal total = BigDecimal.ZERO;
 
     public void add(final Product product) {
         if (items.containsKey(product)) {
             items.replace(product, items.get(product) + 1);
         } else {
-            items.put(product, 1L);
+            items.put(product, 1);
         }
         total = total.add(product.getPrice());
     }
