+++ server/src/main/java/com/eestienergia/fundraiser/domain/ProductCode.java
@@ -2,9 +2,8 @@ package com.eestienergia.fundraiser.domain;
 
 import lombok.Value;
 
-import java.math.BigDecimal;
-
 @Value(staticConstructor = "of")
 public class ProductCode {
-    long code;
+
+    int code;
 }
