+++ server/src/main/java/com/eestienergia/fundraiser/domain/exception/StockIncreaseException.java
@@ -1,7 +1,7 @@
 package com.eestienergia.fundraiser.domain.exception;
 
 public class StockIncreaseException extends RuntimeException {
-    public StockIncreaseException(long amount) {
+    public StockIncreaseException(int amount) {
         super("Cannot increase stock by " + amount);
     }
 }
