+++ server/src/main/java/com/eestienergia/fundraiser/domain/exception/StockUpdateException.java
@@ -1,7 +1,7 @@
 package com.eestienergia.fundraiser.domain.exception;
 
 public class StockUpdateException extends RuntimeException {
-    public StockUpdateException(long amount) {
+    public StockUpdateException(int amount) {
         super("Cannot update stock to " + amount);
     }
 }
